+++ sysconfig.php
@@ -35,6 +35,8 @@ require_once 'inc/head.php';
 require_once 'inc/info_box.php';
 require_once 'vendor/autoload.php';
 
+$crypto = new \Elabftw\Elabftw\Crypto();
+
 $formKey = new \Elabftw\Elabftw\FormKey();
 
 if (strlen(get_config('smtp_username')) == 0) {
@@ -57,7 +59,7 @@ if ($current_version == 'something') {
     <button type='submit' class='submit button'>Update elabftw</button>
     </form>
     </div>
-<?php
+    <?php
 }
 ?>
 
@@ -168,7 +170,7 @@ if ($current_version == 'something') {
 
 <!-- TAB 3 -->
 <div class='divhandle' id='tab3div'>
-    <h3><?php echo _('Universign timestamping configuration'); ?></h3>
+    <h3><?php echo _('Timestamping configuration'); ?></h3>
     <form method='post' action='app/admin-exec.php'>
         <label for='stampshare'><?php echo _('The teams can use the credentials below to timestamp:'); ?></label>
         <select name='stampshare' id='stampshare'>
@@ -179,13 +181,22 @@ if ($current_version == 'something') {
                     if (get_config('stampshare') == 0) { echo " selected='selected'"; } ?>
             ><?php echo _('No'); ?></option>
         </select>
-        <p class='smallgray'><?php echo _('You can control if the teams can use the global Universign account. If set to <em>no</em> the team admin must add login infos in the admin panel.'); ?></p>
+        <p class='smallgray'><?php echo _('You can control if the teams can use the global timestamping account. If set to <em>no</em> the team admin must add login infos in the admin panel.'); ?></p>
+        <p>
+        <label for='stampprovider'><?php echo _('URL for external timestamping service:');?></label>
+        <input type='url' placeholder='https://ws.universign.eu/tsa' value='<?php echo get_config('stampprovider');?>' name='stampprovider' id='stampprovider' />
+        <span class='smallgray'><?php printf(_('This should be the URL used for %sRFC 3161%s-compliant timestamping requests.'), "<a href='https://tools.ietf.org/html/rfc3161'>", "</a>"); ?></span>
+        </p>
+        <p>
+        <label for='stampcert'><?php echo _('Chain of certificates of the external timestamping service:');?></label>
+        <input type='text' placeholder='vendor/universign-tsa-root.pem' value='<?php echo get_config('stampcert');?>' name='stampcert' id='stampcert' />
+        <span class='smallgray'><?php printf(_('This should point to the chain of certificates used by your external timestamping provider to sign the timestamps.%sLocal path relative to eLabFTW installation directory. The file needs to be in %sPEM-encoded (ASCII)%s format!'), "<br>", "<a href='https://en.wikipedia.org/wiki/Privacy-enhanced_Electronic_Mail'>", "</a>"); ?></span>
+        </p>
         <label for='stamplogin'><?php echo _('Login for external timestamping service:'); ?></label>
-        <input type='email' value='<?php echo get_config('stamplogin'); ?>' name='stamplogin' id='stamplogin' />
-        <p class='smallgray'><?php echo _('Must be an email address.'); ?></p>
+        <input type='text' value='<?php echo get_config('stamplogin'); ?>' name='stamplogin' id='stamplogin' />
+        <p class='smallgray'><?php echo _('Login for external timestamping service .'); ?></p>
         <label for='stamppass'><?php echo _('Password for external timestamping service:'); ?></label>
-        <input type='password' value='<?php echo get_config('stamppass'); ?>' name='stamppass' id='stamppass' />
-        <p class='smallgray'><?php echo _("This password will be stored in clear in the database ! Make sure it doesn't open other doors…"); ?></p>
+        <input type='password' value='<?php echo $crypto->decrypt(get_config('stamppass')); ?>' name='stamppass' id='stamppass' />
         <div class='center'>
         <button type='submit' name='submit_config' class='submit button'><?php echo _('Save'); ?></button>
         </div>
@@ -244,7 +255,7 @@ if ($current_version == 'something') {
         </p>
         <p>
         <label for='smtp_password'><?php echo _('SMTP password'); ?></label>
-        <input type='password' value='<?php echo get_config('smtp_password'); ?>' name='smtp_password' id='smtp_password' />
+        <input type='password' value='<?php echo $crypto->decrypt(get_config('smtp_password')); ?>' name='smtp_password' id='smtp_password' />
         </p>
         <div class='center'>
             <button type='submit' name='submit_config' class='submit button'><?php echo _('Save'); ?></button>